New attributes:
'inline=<boolean>': pdf-files can be shown inline in the web-browser.

Bug-fixes

Solved problem that the 'format'-attribute did not accept spaces.
Solved problem that sometimes the from-parameter was not found.
The default locale of the report is the request locale.
The default subreport folder is the same of the report.

Thanks to Achim Breunig